pyproject.toml
[project]
name = "profmatch"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.31.0",
    "email-validator>=2.3.0",
    "fastapi>=0.128.0",
    "google-genai>=1.56.0",
    "greenlet>=3.3.0",
    "httpx>=0.28.1",
    "mcp>=1.25.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pypdf>=6.5.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]
